# Regex: reduce allocation slightly, add tests, code cleanup, add parser comments