--- a/src/Common/src/CoreLib/System/Collections/Generic/ValueListBuilder.cs
+++ b/src/Common/src/CoreLib/System/Collections/Generic/ValueListBuilder.cs
@@ -21,7 +21,16 @@ namespace System.Collections.Generic
             _pos = 0;
         }
 
-        public int Length => _pos;
+        public int Length
+        {
+            get => _pos;
+            set
+            {
+                Debug.Assert(value >= 0);
+                Debug.Assert(value <= _span.Length);
+                _pos = value;
+            }
+        }
 
         public ref T this[int index]
         {