--- a/src/System.Text.RegularExpressions/tests/Regex.Match.Tests.cs
+++ b/src/System.Text.RegularExpressions/tests/Regex.Match.Tests.cs
@@ -291,6 +291,9 @@ namespace System.Text.RegularExpressions.Tests
             // \c
             if (!PlatformDetection.IsFullFramework) // missing fix for #26501
                 yield return new object[] { @"(cat)(\c[*)(dog)", "asdlkcat\u00FFdogiwod", RegexOptions.None, 0, 15, false, string.Empty };
+
+            // Surrogate pairs splitted up into UTF-16 code units.
+            yield return new object[] { @"(\uD82F[\uDCA0-\uDCA3])", "\uD82F\uDCA2", RegexOptions.CultureInvariant, 0, 2, true, "\uD82F\uDCA2" };
         }
 
         [Theory]