- Make RegexParser a ref struct and use ValueListBuilder with Span for the options stack instead of a List<RegexOptions>.
- Add tests for Group.Synchronized & regex comments `#comment` **and other corner cases that weren't covered especially with right to left**.
- Replace some method entry condition checks with Debug Asserts where applicable.
- Code cleanup & simplify code in a few cases (readonly, ToLower on Span instead of single chars, variable declaration moved away from on top of method, some comments replaced with XML comments)
- Add comments to better describe RegexParser behavior.

## Before (master)

|    Method |     Mean |     Error |    StdDev |      Gen 0 | Allocated |
|---------- |---------:|----------:|----------:|-----------:|----------:|
| RegexCtor | 101.8 ms | 0.5352 ms | 0.4179 ms | 56500.0000 | 113.07 MB |
| RegexCtorIgnoreCase | 98.15 ms | 1.913 ms | 2.278 ms | 66312.5000 | 132.64 MB |

## After

|    Method |     Mean |     Error |    StdDev |      Gen 0 | Allocated |
|---------- |---------:|----------:|----------:|-----------:|----------:|
| RegexCtor | 97.78 ms | 0.5337 ms | 0.4731 ms | 53687.5000 | 107.44 MB |
| RegexCtorIgnoreCase | 94.40 ms | 1.030 ms | 0.8600 ms | 62312.5000 | 124.65 MB |

Only minor improvements, I didn't expect anything severe. 5-6% less allocation, ~2-3% better perf.
Contributes towards https://github.com/dotnet/corefx/issues/30507