@jkotas 2018-06-24
What is the `RegexCtor` benchmark doing? I am not able to find it in https://github.com/dotnet/corefx/tree/master/src/System.Text.RegularExpressions/tests/Performance