@ViktorHofer 2018-06-24
I use the following test code https://gist.github.com/ViktorHofer/2b8338fc164ea105f79a456856ff43cb and I run it locally with this script:

```
.\CoreRun ..\..\..\tools\csc.dll /noconfig /optimize /r:System.Private.Corelib.dll /r:System.Runtime.dll /r:System.Runtime.Extensions.dll /r:System.Console.dll /r:System.Text.RegularExpressions.dll /r:System.Collections.dll /r:BenchmarkDotNet.dll /r:BenchmarkDotNet.Core.dll /out:regex-perf.dll regex-perf.cs
.\CoreRun regex-perf.dll
```