@ViktorHofer 2018-06-25
@dotnet-bot test Linux arm64 Release Build
@dotnet-bot test Linux x64 Release Build
@dotnet-bot test Linux-musl x64 Debug Build