@ViktorHofer 2018-06-26
@dotnet-bot test Linux arm Release Build
@dotnet-bot test Tizen armel Debug Build