@MarcoRossignoli 2018-06-28
> We don't recommend using this approach as it involves using internals (csc) and manually referencing the ref assemblies. That said, I documented it a few months ago under advanced inner loop in docs.

Understood, however it's an "alternative" until BDN benchmarking will be fix. Found [doc](https://github.com/dotnet/corefx/blob/a938b5dd0260644eda3ffff0b273d71ccf5a440a/Documentation/building/advanced-inner-loop-testing.md) thank's a lot!