@ViktorHofer 2018-06-28
@danmosemsft @stephentoub do you mind reviewing / approving?