@danmosemsft 2018-06-29
Is there a missing test, based on your last commit here