@ViktorHofer 2018-06-29
No, that's an internal behavior that can't be captured by a test. Not using the locals from the transferred state didn't cause an error but probably would have regressed perf.